Refactor ClickHouse integration and unit tests for atomic updates
- Changed version type from uint32 to uint64 in ClickHouse integration tests and related structures to support nanosecond timestamps.
- Introduced new integration tests for atomic updates ensuring unique versions and data integrity under high concurrency.
- Updated unit tests to reflect changes in the update mechanism, removing ForceUpdate in favor of a standard Update method that leverages timestamp-based versioning.
- Modified client methods to use the new Update method, ensuring proper handling of terminal state transitions without optimistic locking.
- Adjusted mock store implementations and tests to align with the removal of ForceUpdate, ensuring consistency across the codebase.
- Updated JSON configuration to reflect changes in prerequisites for deployment steps.
- Enhanced dynamic migrations to accommodate the new versioning scheme.

+++ slippy/dynamic_migrations.go
@@ -278,8 +278,8 @@ func (m *DynamicMigrationManager) generateBaseTableMigration() clickhousemigrato
 				-- VersionedCollapsingMergeTree columns
 				-- sign: 1 = active row, -1 = cancelled row (for updates)
 				sign Int8 DEFAULT 1,
-				-- version: monotonically increasing version number for each correlation_id
-				version UInt32 DEFAULT 1,
+				-- version: nanosecond timestamp for unique version across concurrent writers
+				version UInt64 DEFAULT 1,
 
 				-- Bloom filter indexes
 				INDEX idx_repository repository TYPE bloom_filter GRANULARITY 1,
